requirements.txt
appJar==0.93.0
astroid==2.1.0
beautifulsoup4==4.7.1
colorama==0.4.1
cycler==0.10.0
decrunch==0.3.1
fsb5==1.0
imutils==0.5.2
isort==4.3.4
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
lz4==2.1.6
matplotlib==3.0.3
mccabe==0.6.1
mss==4.0.2
numpy==1.16.3
opencv-python==3.4.5.20
pandas==0.24.2
Pillow==5.3.0
PyAutoGUI==0.9.42
PyGetWindow==0.0.4
pylint==2.2.2
PyMsgBox==1.0.6
pyparsing==2.4.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
PyRect==0.1.4
PyScreeze==0.1.20
pytesseract==0.2.6
python-dateutil==2.8.0
PyTweening==1.0.3
pytz==2019.1
pywin32==224
Reindent==0.1.1
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
soupsieve==1.7.3
tesseract==0.1.3
unitypack==0.9.0
wrapt==1.10.11
